+++ src/services/authServices.ts
@@ -1,4 +1,4 @@
-import { httpService } from "./http.service";
+import { httpService } from "./httpService";
 import { IUser } from "@/app/login/login.type";
 
 class AuthServices {
